+++ src/replace.ts
@@ -239,7 +239,7 @@ const addFilesSync = (p: Pack, files: string[]) => {
         file: path.resolve(p.cwd, file.slice(1)),
         sync: true,
         noResume: true,
-        onentry: entry => p.add(entry),
+        onReadEntry: entry => p.add(entry),
       })
     } else {
       p.add(file)
@@ -258,7 +258,7 @@ const addFilesAsync = async (
       await list({
         file: path.resolve(String(p.cwd), file.slice(1)),
         noResume: true,
-        onentry: entry => p.add(entry),
+        onReadEntry: entry => p.add(entry),
       })
     } else {
       p.add(file)